Add theme change functionality to dashboard
This commit adds the ability to change themes on the dashboard. Main changes include adding a "theme-change" dependency to the package and configuring it in the tailwind config file. The theme change is then activated by a button in the dashboard file. Additional themes were also added in application.css. This allows more customization and improves user experience.