Add the International Space Station (ISS) features
This commit adds the ability to display the location and population of the ISS on the dashboard. A reference to the IssService is included in the HomeController which makes calls to get the location and population of the ISS. The responses are modeled in PeopleItem and IssLocationResponse respectively. A new file, iss-fragments.html, has been created to hold the HTML fragments for displaying the ISS data. The index.html file has been significantly modified to integrate these changes.

This update is necessary because the requirement was made to track and show ISS real-time data as part of the dashboard's main features.