Refactor fragments and enhance UI with user location and Hobis quotes
Renamed "iss-fragments" to "fragments" to clarify that these are general fragments used in the project, not only related to the ISS. Enhanced the UI by replacing the grid layout with flex layout for displaying latitude and longitude details. Added id attributes to latitude and longitude spans for ISS and user for easier access. Removed unnecessary comments. Included a new function to calculate the distance between the user and ISS and display it on the page. Also, added a new HxRequest named "hobis" that pulls up random quotes from "Hobis says.txt" and displays them on the main page.